Add new ignoreFiles array to device plugin manager
All manager tests now pass on Windows. The device plugin now has a new
ignoreFiles string array which specifies the names of files to be
ignored when looking for new device plugin sockets to register. By
default, the server socket itself and the kubelet internal checkpoint
are to be ignored.